requirements.txt
torch==2.5.1
torchvision==0.20.1
trl==0.13.0
transformers==4.50.3
tqdm @ file:///croot/tqdm_1724853939799/work
pillow==10.4.0
PyYAML==6.0.1
numpy==1.26.4
datasets==3.2.0
scikit-learn==1.5.2